Added HelpPanel component to PenteGameWindow.
This component dynamically loads its content from MultiGameFlexClient/src/main/webapp/help-gente.html the idea being that we could reuse the same
html file for other purposes. In reality the html that can be included in this file is very limited. Flash can only understand a small subset of html
tags and does not render the same way as other browsers. This would work better in an AIR application since the AIR runtime includes Webkit.