Closes #31.
The null pointer exception was caused by setting the text property of the GraphicalAlert before adding it to the stage.

The setter for this property asigned the argument to a text fields text property. However since the GraphicalAlert had not yet been added to the stage this text field had not been created. Hense the null pointer exception.

The solution is to bind the value of the text field to the property. This allows for defining the property before adding the component to the stage and once the text field is created it will initialize with the saved value.

In addition the icon displayed has been set to "positive" since the user has won the stage.