Closes #24. Some code had already been written in ManantialesGameController.initGrid to remove the tokens from each store when the grid initializes. However this executes before the player gets the turn and the token stores are created by a state change after the player gets their turn.
An initialization flag has been added to the ManantialesTokenStore and the removing of tokens put in ManantialesGameController.initTurn.

In addition on creation complete ManantialesTokenStore.reset is called. This recreated the store with the full 6 tokens overwritting the previous change. Recreation of the tokens is not necessary here and has been removed.