Fixed bug in move animation in gente after build with flex 4.
Adding a move to the move viewer causes the accordion component to execute animations. This can cause the move animation not to be rendered (a bit strange). This did not happen when compiled with flex 3.

The solution implemented in this commit is to add the new move to the viewer after the animation finishes, which is actually more intuitive from a UI perspective.